refactor(daemon,tui,herdr): move inline tests to child modules
